Code reorg to put more cache code together; fix problems i discover in the process
