+++ include/dpu_offload_group_cache.h
@@ -6,6 +6,8 @@
 
 #ifndef DPU_OFFLOAD_GROUP_CACHE_H_
 
+#define GROUP_SIZE_UNKNOWN (-1)
+
 /* GROUPS_CACHE_INIT initializes the cache that holds information about all the groups */
 #define GROUPS_CACHE_INIT(_cache)                                                            \
     do                                                                                       \
@@ -60,6 +62,18 @@
         void *_gp_cache = _cache[_gp_id];                                    \
         DYN_ARRAY_ALLOC((dyn_array_t *)_gp_cache, 2048, peer_cache_entry_t); \
     } while (0)
+
+/**
+ * @brief Checks whether a rank in a given group is in the cache.
+ *
+ * @param cache Pointer to the target cache to query
+ * @param gp_uid Target group's UID
+ * @param rank_id Target rank in the group
+ * @param group_size Group size (can be GROUP_SIZE_UNKNOWN)
+ * @return true
+ * @return false
+ */
+bool is_in_cache(cache_t *cache, group_uid_t gp_uid, int64_t rank_id, int64_t group_size);
 
 /**
  * @brief This function populated the cache's lookup tables. It assumes the cache is
@@ -404,5 +418,53 @@ dpu_offload_status_t update_topology_data(offloading_engine_t *engine, group_cac
  * @return dpu_offload_status_t
  */
 dpu_offload_status_t host_add_local_rank_to_cache(offloading_engine_t *engine, rank_info_t *rank_info);
+
+
+/**
+ * @brief Get the dpu ID by host rank object. That ID can then be used to look up the corresponding endpoint.
+ *
+ * @param[in] engine Offloading engine for the query
+ * @param[in] gp_uid Target group's UID
+ * @param[in] rank Target rank in the group
+ * @param[in] sp_idx In case of multiple service processes per host, index of the target shadow service process for the group/rank
+ * @param[out] cb Associated callback. If the event completes right away, the callback is still being invoked. The user is in charge of returning the object related to the request.
+ * @return dpu_offload_status_t
+ *
+ * Example:
+ *      - To issue the request for the first service process attached to the group/rank `gp_id` and `rank_id`
+ *          get_cache_entry_by_group_rank(offload_engine, gp_id, rank_id, 0, my_completion_cb);
+ *      - Completion callback example:
+ *          void my_completion_cb(void *data)
+ *          {
+ *              assert(data);
+ *              cache_entry_request_t *cache_entry_req = (cache_entry_request_t*)data;
+ *              assert(cache_entry_req->offload_engine);
+ *              offloading_engine_t *engine = (offloading_engine_t*)cache_entry_req->offload_engine;
+ *              ucp_ep_h target_sp_ep = NULL;
+ *              execution_context_t *target_sp_econtext = NULL;
+ *              uint64_t notif_dest_id;
+ *              get_sp_ep_by_id(engine, cache_entry_req->target_sp_idx, &target_sp_econtext, &target_sp_econtext, &notif_dest_id);
+ *              assert(target_sp_ep == NULL);
+ *              DYN_LIST_RETURN(engine->free_cache_entry_requests, cache_entry_req, item);
+ *          }
+ */
+dpu_offload_status_t get_cache_entry_by_group_rank(offloading_engine_t *engine, group_uid_t gp_uid, int64_t rank, int64_t sp_idx, request_compl_cb_t cb);
+
+/**
+ * @brief Get the global service process ID (not the group service process global ID) that is associated with a specific rank in a group.
+ * The global ID can then be used to look up the corresponding endpoint, for example to issue a XGVMI operation.
+ * The global identifier is the identifier set a startup time; it is not a group-level identifier.
+ *
+ * @param[in] engine Offloading engine for the query
+ * @param[in] gp_uid Target group's UID
+ * @param[in] rank Target rank in the group
+ * @param[in] dpu_idx In case of multiple service processes per host, index of the target shadow service process for the group/rank
+ * @param[out] dpu_id Resulting service process identifier
+ * @param[out] ev Associated event. If NULL, the DPU identifier is available right away. If not, it is required to call the function again once the event has completed. The caller is in charge of returning the event after completion. The event cannot be added to any list since it is already put on a list.
+ * @return dpu_offload_status_t
+ */
+dpu_offload_status_t get_sp_id_by_group_rank(offloading_engine_t *engine, group_uid_t gp_uid, int64_t rank, int64_t sp_idx, int64_t *sp_id, dpu_offload_event_t **ev);
+
+void display_group_cache(cache_t *cache, group_uid_t gp_uid);
 
 #endif // DPU_OFFLOAD_GROUP_CACHE_H_
