+++ src/dpu_offload_group_cache.c
@@ -9,19 +9,38 @@
 
 #include "dpu_offload_types.h"
 #include "dpu_offload_debug.h"
+#include "dpu_offload_group_cache.h"
+#include "dpu_offload_event_channels.h"
 
 // Forward declarations
 static dpu_offload_status_t do_populate_group_cache_lookup_table(offloading_engine_t *engine, group_cache_t *gp_cache);
 dpu_offload_status_t offload_engine_progress(offloading_engine_t *engine);
+dpu_offload_status_t do_send_cache_entry_request(execution_context_t *econtext, ucp_ep_h ep, uint64_t dest_id, rank_info_t *requested_peer, dpu_offload_event_t *ev);
+dpu_offload_status_t send_cache_entry_request(execution_context_t *econtext, ucp_ep_h ep, uint64_t dest_id, rank_info_t *requested_peer, dpu_offload_event_t **ev);
 
+#if 0
 #define PROGRESS_UNTIL_GROUP_CACHE_FULLY_POPULATED(__engine, __gp_uid)  \
     do                                                                  \
     {                                                                   \
-        while (!group_cache_populated((__engine), (__gp_uid)))          \
+        fprintf(stderr, "[DBG] Waiting for cache 0x%x to be populated\n", __gp_uid);   \
+        group_cache_t *__gpcache = NULL;\
+        __gpcache = GET_GROUP_CACHE(&(engine->procs_cache), __gp_uid);\
+        assert(__gpcache);\
+        while (__gpcache->group_size == 0 || !group_cache_populated((__engine), (__gp_uid)))          \
         {                                                               \
             offload_engine_progress((__engine));                        \
         }                                                               \
+        fprintf(stderr, "[DBG] Cache populated (size: %ld)\n", __gpcache->group_size);                     \
     } while (0)
+#endif
+
+bool is_in_cache(cache_t *cache, group_uid_t gp_uid, int64_t rank_id, int64_t group_size)
+{
+    peer_cache_entry_t *entry = GET_GROUP_RANK_CACHE_ENTRY(cache, gp_uid, rank_id, group_size);
+    if (entry == NULL)
+        return false;
+    return (entry->set);
+}
 
 bool group_cache_populated(offloading_engine_t *engine, group_uid_t gp_uid)
 {
@@ -36,6 +55,32 @@ bool group_cache_populated(offloading_engine_t *engine, group_uid_t gp_uid)
     return false;
 }
 
+void display_group_cache(cache_t *cache, group_uid_t gp_uid)
+{
+    size_t i = 0;
+    size_t idx = 0;
+    group_cache_t *gp_cache = NULL;
+    gp_cache = GET_GROUP_CACHE(cache, gp_uid);
+    fprintf(stderr, "Content of cache for group 0x%x\n", gp_uid);
+    fprintf(stderr, "-> group size: %ld\n", gp_cache->group_size);
+    fprintf(stderr, "-> n_local_rank: %ld\n", gp_cache->n_local_ranks);
+    fprintf(stderr, "-> n_local_ranks_populated: %ld\n", gp_cache->n_local_ranks_populated);
+    fprintf(stderr, "-> num_local_entries: %ld\n", gp_cache->num_local_entries);
+    fprintf(stderr, "-> sent_to_host: %d\n\n", gp_cache->sent_to_host);
+
+    while (i < gp_cache->group_size)
+    {
+        peer_cache_entry_t *entry = GET_GROUPRANK_CACHE_ENTRY(cache, gp_uid, idx);
+        if (entry->set)
+        {
+            fprintf(stderr, "Rank %" PRId64 " host: 0x%lx\n", entry->peer.proc_info.group_rank, entry->peer.host_info);
+            assert(idx == entry->peer.proc_info.group_rank);
+            i++;
+        }
+        idx++;
+    }
+}
+
 dpu_offload_status_t
 get_global_sp_id_by_group(offloading_engine_t *engine,
                           group_uid_t gp_uid,
@@ -48,7 +93,9 @@ get_global_sp_id_by_group(offloading_engine_t *engine,
     if (!engine->on_dpu)
         return DO_ERROR;
 
-    PROGRESS_UNTIL_GROUP_CACHE_FULLY_POPULATED(engine, gp_uid);
+    INFO_MSG("[DBG]");
+
+    //PROGRESS_UNTIL_GROUP_CACHE_FULLY_POPULATED(engine, gp_uid);
     gp_cache = GET_GROUP_CACHE(&(engine->procs_cache), gp_uid);
     assert(gp_cache);
     for (i = 0; i < gp_cache->n_sps; i++)
@@ -79,8 +126,10 @@ get_local_sp_id_by_group(offloading_engine_t *engine,
     sp_cache_data_t *sp_data = NULL;
     group_cache_t *gp_cache = NULL;
 
+    INFO_MSG("[DBG]");
+
     assert(engine);
-    PROGRESS_UNTIL_GROUP_CACHE_FULLY_POPULATED(engine, gp_uid);
+    //PROGRESS_UNTIL_GROUP_CACHE_FULLY_POPULATED(engine, gp_uid);
     gp_cache = GET_GROUP_CACHE(&(engine->procs_cache), gp_uid);
     assert(gp_cache);
 
@@ -113,8 +162,10 @@ get_host_idx_by_group(offloading_engine_t *engine,
     host_uid_t my_host_uid;
     size_t i;
 
+    INFO_MSG("[DBG]");
+
     assert(engine);
-    PROGRESS_UNTIL_GROUP_CACHE_FULLY_POPULATED(engine, group_uid);
+    //PROGRESS_UNTIL_GROUP_CACHE_FULLY_POPULATED(engine, group_uid);
     my_host_uid = engine->config->local_service_proc.host_uid;
     gp_cache = GET_GROUP_CACHE(&(engine->procs_cache), group_uid);
     assert(gp_cache);
@@ -158,8 +209,10 @@ get_num_sps_by_group_host_idx(offloading_engine_t *engine,
     host_cache_data_t *host_data = NULL;
     group_cache_t *gp_cache = NULL;
 
+    INFO_MSG("[DBG]");
+
     assert(engine);
-    PROGRESS_UNTIL_GROUP_CACHE_FULLY_POPULATED(engine, group_uid);
+    //PROGRESS_UNTIL_GROUP_CACHE_FULLY_POPULATED(engine, group_uid);
     gp_cache = GET_GROUP_CACHE(&(engine->procs_cache), group_uid);
     assert(gp_cache);
 
@@ -195,8 +248,10 @@ get_num_ranks_for_group_sp(offloading_engine_t *engine,
     sp_cache_data_t *sp_info = NULL;
     group_cache_t *gp_cache = NULL;
 
+    INFO_MSG("[DBG]");
+
     assert(engine);
-    PROGRESS_UNTIL_GROUP_CACHE_FULLY_POPULATED(engine, group_uid);
+    //PROGRESS_UNTIL_GROUP_CACHE_FULLY_POPULATED(engine, group_uid);
     gp_cache = GET_GROUP_CACHE(&(engine->procs_cache), group_uid);
     assert(gp_cache);
 
@@ -232,8 +287,10 @@ get_num_ranks_for_group_host_local_sp(offloading_engine_t *engine,
     host_cache_data_t *host_data = NULL;
     sp_cache_data_t **sp_data_ptr = NULL;
 
+    INFO_MSG("[DBG]");
+
     assert(engine);
-    PROGRESS_UNTIL_GROUP_CACHE_FULLY_POPULATED(engine, group_uid);
+    //PROGRESS_UNTIL_GROUP_CACHE_FULLY_POPULATED(engine, group_uid);
     gp_cache = GET_GROUP_CACHE(&(engine->procs_cache), group_uid);
     assert(gp_cache);
 
@@ -275,8 +332,10 @@ get_num_ranks_for_group_host_idx(offloading_engine_t *engine,
     group_cache_t *gp_cache = NULL;
     host_cache_data_t *host_data = NULL;
 
+    INFO_MSG("[DBG]");
+
     assert(engine);
-    PROGRESS_UNTIL_GROUP_CACHE_FULLY_POPULATED(engine, group_uid);
+    //PROGRESS_UNTIL_GROUP_CACHE_FULLY_POPULATED(engine, group_uid);
     gp_cache = GET_GROUP_CACHE(&(engine->procs_cache), group_uid);
     assert(gp_cache);
 
@@ -307,15 +366,17 @@ get_rank_idx_by_group_host_idx(offloading_engine_t *engine,
                                group_uid_t group_uid,
                                size_t host_idx,
                                int64_t rank,
-                               int64_t *idx)
+                               uint64_t *idx)
 {
     group_cache_t *gp_cache = NULL;
     host_info_t **host_info_ptr = NULL;
     host_cache_data_t *host_data = NULL;
     size_t i, rank_index = 0;
 
+    INFO_MSG("[DBG]");
+
     assert(engine);
-    PROGRESS_UNTIL_GROUP_CACHE_FULLY_POPULATED(engine, group_uid);
+    //PROGRESS_UNTIL_GROUP_CACHE_FULLY_POPULATED(engine, group_uid);
     gp_cache = GET_GROUP_CACHE(&(engine->procs_cache), group_uid);
     assert(gp_cache);
 
@@ -367,8 +428,10 @@ get_rank_idx_by_group_sp_id(offloading_engine_t *engine,
     group_cache_t *gp_cache = NULL;
     size_t rank_index;
 
+    INFO_MSG("[DBG]");
+
     assert(engine);
-    PROGRESS_UNTIL_GROUP_CACHE_FULLY_POPULATED(engine, group_uid);
+    //PROGRESS_UNTIL_GROUP_CACHE_FULLY_POPULATED(engine, group_uid);
     gp_cache = GET_GROUP_CACHE(&(engine->procs_cache), group_uid);
     assert(gp_cache);
 
@@ -417,8 +480,10 @@ get_all_sps_by_group_host_idx(offloading_engine_t *engine,
     host_info_t **host_info_ptr = NULL;
     host_cache_data_t *host_data = NULL;
 
+    INFO_MSG("[DBG]");
+
     assert(engine);
-    PROGRESS_UNTIL_GROUP_CACHE_FULLY_POPULATED(engine, group_uid);
+    //PROGRESS_UNTIL_GROUP_CACHE_FULLY_POPULATED(engine, group_uid);
     gp_cache = GET_GROUP_CACHE(&(engine->procs_cache), group_uid);
     assert(gp_cache);
 
@@ -452,9 +517,11 @@ get_all_hosts_by_group(offloading_engine_t *engine,
     group_cache_t *gp_cache = NULL;
 
     assert(engine);
-    PROGRESS_UNTIL_GROUP_CACHE_FULLY_POPULATED(engine, group_uid);
+    //PROGRESS_UNTIL_GROUP_CACHE_FULLY_POPULATED(engine, group_uid);
     gp_cache = GET_GROUP_CACHE(&(engine->procs_cache), group_uid);
     assert(gp_cache);
+
+    INFO_MSG("[DBG]");
 
     // We still assume this function can only be called when the cache is fully populated
     // but the lookup tables, when on the host, are not necessarily ready
@@ -484,8 +551,10 @@ get_all_ranks_by_group_sp_gid(offloading_engine_t *engine,
     remote_service_proc_info_t **ptr = NULL;
     sp_cache_data_t *sp_data = NULL;
 
+    INFO_MSG("[DBG]");
+
     assert(engine);
-    PROGRESS_UNTIL_GROUP_CACHE_FULLY_POPULATED(engine, group_uid);
+    //PROGRESS_UNTIL_GROUP_CACHE_FULLY_POPULATED(engine, group_uid);
     gp_cache = GET_GROUP_CACHE(&(engine->procs_cache), group_uid);
     assert(gp_cache);
 
@@ -525,8 +594,10 @@ get_all_ranks_by_group_sp_lid(offloading_engine_t *engine,
     host_cache_data_t *host_data = NULL;
     sp_cache_data_t **sp_data_ptr = NULL;
 
+    INFO_MSG("[DBG]");
+
     assert(engine);
-    PROGRESS_UNTIL_GROUP_CACHE_FULLY_POPULATED(engine, group_uid);
+    //PROGRESS_UNTIL_GROUP_CACHE_FULLY_POPULATED(engine, group_uid);
     gp_cache = GET_GROUP_CACHE(&(engine->procs_cache), group_uid);
     assert(gp_cache);
 
@@ -568,8 +639,10 @@ get_nth_sp_by_group_host_idx(offloading_engine_t *engine,
     host_cache_data_t *host_data = NULL;
     sp_cache_data_t **sp_data_ptr = NULL;
 
+    INFO_MSG("[DBG]");
+
     assert(engine);
-    PROGRESS_UNTIL_GROUP_CACHE_FULLY_POPULATED(engine, group_uid);
+    //PROGRESS_UNTIL_GROUP_CACHE_FULLY_POPULATED(engine, group_uid);
     gp_cache = GET_GROUP_CACHE(&(engine->procs_cache), group_uid);
     assert(gp_cache);
 
@@ -609,14 +682,16 @@ dpu_offload_status_t get_sp_group_gid(offloading_engine_t *engine,
     group_cache_t *gp_cache = NULL;
     size_t sp_gp_idx;
 
+    INFO_MSG("[DBG]");
+
     assert(engine);
-    PROGRESS_UNTIL_GROUP_CACHE_FULLY_POPULATED(engine, group_uid);
+    //PROGRESS_UNTIL_GROUP_CACHE_FULLY_POPULATED(engine, group_uid);
+    INFO_MSG("[DBG] Getting SP group GID for group 0x%x and SP %ld\n",
+             group_uid, sp_gid);
     gp_cache = GET_GROUP_CACHE(&(engine->procs_cache), group_uid);
     assert(gp_cache);
 
-    // We still assume this function can only be called when the cache is fully populated
-    // but the lookup tables, when on the host, are not necessarily ready
-    if (!engine->on_dpu && !gp_cache->lookup_tables_populated)
+    if (!gp_cache->lookup_tables_populated)
     {
         dpu_offload_status_t rc;
         rc = do_populate_group_cache_lookup_table(engine, gp_cache);
@@ -626,9 +701,19 @@ dpu_offload_status_t get_sp_group_gid(offloading_engine_t *engine,
         }
     }
 
+#if !NDEBUG
+    if (!gp_cache->sp_array_initialized && !group_cache_populated(engine, group_uid))
+    {
+        ERR_MSG("Group cache lookup tables are not created and the group cache incomplete");
+        return DO_ERROR;
+    }
+#endif
+
+    assert(gp_cache->sp_array_initialized);
     for (sp_gp_idx = 0; sp_gp_idx < gp_cache->n_sps; sp_gp_idx++)
     {
         remote_service_proc_info_t **sp_data = NULL;
+        INFO_MSG("[DBG] CM sp_gp_idx: %ld, n_sps: %ld, capacity: %ld\n", sp_gp_idx, gp_cache->n_sps, (&(gp_cache->sps))->capacity);
         sp_data = DYN_ARRAY_GET_ELT(&(gp_cache->sps), sp_gp_idx, remote_service_proc_info_t *);
         assert(sp_data);
 
@@ -646,6 +731,7 @@ static void
 populate_sp_ranks(offloading_engine_t *engine, group_cache_t *gp_cache, sp_cache_data_t *sp_data)
 {
     size_t i = 0, idx = 0;
+    INFO_MSG("[DBG]");
     DYN_ARRAY_ALLOC(&(sp_data->ranks),
                     gp_cache->group_size,
                     peer_cache_entry_t *);
@@ -675,6 +761,7 @@ static void
 populate_host_sps(group_cache_t *gp_cache, host_cache_data_t *host_data)
 {
     size_t i = 0, idx = 0;
+    INFO_MSG("[DBG]");
     DYN_ARRAY_ALLOC(&(host_data->sps),
                     gp_cache->group_size,
                     sp_cache_data_t *);
@@ -701,6 +788,7 @@ do_populate_group_cache_lookup_table(offloading_engine_t *engine, group_cache_t 
 {
     size_t i, idx = 0;
 
+    INFO_MSG("[DBG]");
     assert(engine);
     assert(gp_cache);
 
@@ -711,6 +799,7 @@ do_populate_group_cache_lookup_table(offloading_engine_t *engine, group_cache_t 
     assert(gp_cache->n_sps);
     if (gp_cache->sp_array_initialized == false)
     {
+        INFO_MSG("[DBG] YAY!!!!\n");
         DYN_ARRAY_ALLOC(&(gp_cache->sps),
                         gp_cache->n_sps,
                         remote_service_proc_info_t *);
@@ -792,6 +881,7 @@ dpu_offload_status_t
 populate_group_cache_lookup_table(offloading_engine_t *engine,
                                   group_cache_t *gp_cache)
 {
+    INFO_MSG("[DBG]");
     assert(gp_cache);
     assert(group_cache_populated(engine, gp_cache->group_uid));
     return do_populate_group_cache_lookup_table(engine, gp_cache);
@@ -804,6 +894,7 @@ update_topology_data(offloading_engine_t *engine, group_cache_t *gp_cache, int64
     sp_cache_data_t *sp_data = NULL;
     host_cache_data_t *host_data = NULL;
 
+    INFO_MSG("[DBG]");
     assert(engine);
     assert(gp_cache);
 
@@ -905,6 +996,7 @@ dpu_offload_status_t
 host_add_local_rank_to_cache(offloading_engine_t *engine, rank_info_t *rank_info)
 {
     dpu_offload_state_t ret;
+    INFO_MSG("[DBG]");
     peer_cache_entry_t *cache_entry = GET_GROUP_RANK_CACHE_ENTRY(&(engine->procs_cache),
                                                                  rank_info->group_uid,
                                                                  rank_info->group_rank,
@@ -927,4 +1019,177 @@ host_add_local_rank_to_cache(offloading_engine_t *engine, rank_info_t *rank_info
     CHECK_ERR_RETURN((ret != DO_SUCCESS), DO_ERROR, "update_topology_data() failed");
 
     return DO_SUCCESS;
+}
+
+static dpu_offload_status_t
+do_get_cache_entry_by_group_rank(offloading_engine_t *engine,
+                                 group_uid_t gp_uid,
+                                 int64_t rank,
+                                 int64_t sp_idx,
+                                 request_compl_cb_t cb,
+                                 int64_t *sp_global_id,
+                                 dpu_offload_event_t **ev)
+{
+    if (ev != NULL && cb != NULL)
+    {
+        ERR_MSG("%s(): both the event and the callback are defined, impossible to understand the context", __func__);
+        return DO_ERROR;
+    }
+
+    // If the event is defined, the dpu_id must also be defined, they go in pairs
+    if (ev != NULL)
+        assert(sp_global_id);
+
+    if (is_in_cache(&(engine->procs_cache), gp_uid, rank, -1))
+    {
+        // The cache has the data
+        peer_cache_entry_t *cache_entry = GET_GROUP_RANK_CACHE_ENTRY(&(engine->procs_cache), gp_uid, rank, GROUP_SIZE_UNKNOWN);
+        DBG("%" PRId64 " from group 0x%x is in the cache, service proc ID = %" PRId64, rank, gp_uid, cache_entry->shadow_service_procs[sp_idx]);
+        if (ev != NULL)
+        {
+            *ev = NULL;
+            *sp_global_id = cache_entry->shadow_service_procs[sp_idx];
+        }
+        return DO_SUCCESS;
+    }
+
+#if !NDEBUG
+    // With the current implementation, it should always be in the cache
+    WARN_MSG("rank %" PRId64 " from group 0x%x is not in the cache", rank, gp_uid);
+    display_group_cache(&(engine->procs_cache), gp_uid);
+    assert(0);
+#endif
+
+    // If the element is not in the cache, we send a request to the service process to get the data.
+    // At the moment we assume by design that the cache is fully populated early on so we should never
+    // face this situation. We keep the code in case we relax the assumptions later so we won't have
+    // to start from scratch.
+    abort();
+
+    // The cache does not have the data. We sent a request to get the data.
+    // The caller is in charge of calling the function after completion to actually get the data
+    rank_info_t rank_data;
+    RESET_RANK_INFO(&rank_data);
+    rank_data.group_uid = gp_uid;
+    rank_data.group_rank = rank;
+
+    // Create the local event so we can know when the cache entry has been received
+    dpu_offload_event_t *cache_entry_updated_ev;
+    peer_cache_entry_t *cache_entry = GET_GROUP_RANK_CACHE_ENTRY(&(engine->procs_cache), gp_uid, rank, GROUP_SIZE_UNKNOWN);
+    dpu_offload_status_t rc = event_get(engine->self_econtext->event_channels, NULL, &cache_entry_updated_ev);
+    CHECK_ERR_RETURN((rc), DO_ERROR, "event_get() failed");
+    if (!cache_entry->events_initialized)
+    {
+        SIMPLE_LIST_INIT(&(cache_entry->events));
+        cache_entry->events_initialized = true;
+    }
+    EVENT_HDR_TYPE(cache_entry_updated_ev) = META_EVENT_TYPE;
+    // We just queue a local event on the list for the cache entry to track what is being done in the
+    // context of that entry
+    SIMPLE_LIST_PREPEND(&(cache_entry->events), &(cache_entry_updated_ev->item));
+    DBG("Cache entry %p for gp/rank 0x%x/%" PRIu64 " now has %ld update events",
+        cache_entry, gp_uid, rank, SIMPLE_LIST_LENGTH(&(cache_entry->events)));
+    if (ev != NULL)
+    {
+        // If the calling function is expecting an event back, no need for anything other than
+        // make sure we return the event
+        *ev = cache_entry_updated_ev;
+    }
+    if (cb != NULL)
+    {
+        // If the calling function specified a callback, the event needs to be hidden from the
+        // caller and the callback will be invoked upon completion. So we need to put the event
+        // on the list of ongoing events. Make sure to never return the event to the caller to
+        // prevent the case where the event could be put on two different lists (which leads
+        // to memory corruptions).
+        cache_entry_request_t *request_data;
+        DYN_LIST_GET(engine->free_cache_entry_requests, cache_entry_request_t, item, request_data);
+        assert(request_data);
+        request_data->gp_uid = gp_uid;
+        request_data->rank = rank;
+        request_data->target_sp_idx = sp_idx;
+        request_data->offload_engine = engine;
+        cache_entry_updated_ev->ctx.completion_cb = cb;
+        cache_entry_updated_ev->ctx.completion_cb_ctx = (void *)request_data;
+        assert(0); // FIXME: events cannot be on two lists
+        // ucs_list_add_tail(&(engine->self_econtext->ongoing_events), &(cache_entry_updated_ev->item));
+    }
+
+    if (engine->on_dpu == true)
+    {
+        // If we are on a DPU, we need to send a request to all known DPUs
+        // To track completion, we get an event from the execution context used for the
+        // first DPU.
+        size_t i;
+        dpu_offload_status_t rc;
+        dpu_offload_event_t *metaev = NULL;
+        execution_context_t *meta_econtext = NULL;
+
+        for (i = 0; i < engine->num_service_procs; i++)
+        {
+            remote_service_proc_info_t *sp;
+            sp = DYN_ARRAY_GET_ELT(&(engine->service_procs), i, remote_service_proc_info_t);
+            assert(sp);
+            if (sp != NULL && sp->ep != NULL && sp->init_params.conn_params != NULL)
+            {
+                execution_context_t *econtext = ECONTEXT_FOR_SERVICE_PROC_COMMUNICATION(engine, i);
+                CHECK_ERR_RETURN((econtext == NULL), DO_ERROR, "unable to get execution context to communicate with service process #%ld", i);
+                uint64_t global_sp_id = LOCAL_ID_TO_GLOBAL(econtext, i);
+                DBG("Sending cache entry request to service process #%ld (econtext: %p, scope_id: %d)",
+                    global_sp_id,
+                    econtext,
+                    econtext->scope_id);
+
+                if (metaev == NULL)
+                {
+                    meta_econtext = econtext;
+                    rc = event_get(meta_econtext->event_channels, NULL, &metaev);
+                    CHECK_ERR_RETURN((rc), DO_ERROR, "get_event() failed");
+                    EVENT_HDR_TYPE(metaev) = META_EVENT_TYPE;
+                }
+
+                ucp_ep_h dpu_ep = sp->ep;
+                dpu_offload_event_t *subev;
+                rc = event_get(econtext->event_channels, NULL, &subev);
+                CHECK_ERR_RETURN((rc), DO_ERROR, "event_get() failed");
+                subev->is_subevent = true;
+                rc = do_send_cache_entry_request(econtext, dpu_ep, i, &rank_data, subev);
+                CHECK_ERR_RETURN((rc), DO_ERROR, "send_cache_entry_request() failed");
+                DBG("Sub-event for sending cache to DPU %ld: %p", global_sp_id, subev);
+                if (subev != NULL)
+                {
+                    // If the event did not complete right away, we add it as a sub-event to the meta-event so we can track everything
+                    QUEUE_SUBEVENT(metaev, subev);
+                }
+            }
+        }
+        if (metaev)
+        {
+            assert(meta_econtext);
+            if (!event_completed(metaev))
+                QUEUE_EVENT(metaev);
+            else
+                event_return(&metaev);
+        }
+    }
+    else
+    {
+        // If we are on the host, we need to send a request to our first shadow DPU
+        DBG("Sending request for cache entry...");
+        execution_context_t *econtext = engine->client;
+        return send_cache_entry_request(econtext, GET_SERVER_EP(econtext), econtext->client->server_id, &rank_data, ev);
+    }
+    return DO_SUCCESS;
+}
+
+dpu_offload_status_t get_cache_entry_by_group_rank(offloading_engine_t *engine, group_uid_t gp_uid, int64_t rank, int64_t sp_idx, request_compl_cb_t cb)
+{
+    //PROGRESS_UNTIL_GROUP_CACHE_FULLY_POPULATED(engine, gp_uid);
+    return do_get_cache_entry_by_group_rank(engine, gp_uid, rank, sp_idx, cb, NULL, NULL);
+}
+
+dpu_offload_status_t get_sp_id_by_group_rank(offloading_engine_t *engine, group_uid_t gp_uid, int64_t rank, int64_t sp_idx, int64_t *sp_id, dpu_offload_event_t **ev)
+{
+    //PROGRESS_UNTIL_GROUP_CACHE_FULLY_POPULATED(engine, gp_uid);
+    return do_get_cache_entry_by_group_rank(engine, gp_uid, rank, sp_idx, NULL, sp_id, ev);
 }
