+++ src/dpu_offload_utils.c
@@ -13,6 +13,7 @@
 #include <unistd.h>
 
 #include "dpu_offload_types.h"
+#include "dpu_offload_service_daemon.h"
 #include "dpu_offload_mem_mgt.h"
 #include "dpu_offload_debug.h"
 #include "dpu_offload_event_channels.h"
@@ -30,19 +31,9 @@ execution_context_t *get_server_servicing_host(offloading_engine_t *engine);
 
 const char *config_file_version_token = "Format version:";
 
-#define GROUP_SIZE_UNKNOWN (-1)
-
 /*************************************/
 /* FUNCTIONS RELATED TO GROUPS/RANKS */
 /*************************************/
-
-bool is_in_cache(cache_t *cache, group_uid_t gp_uid, int64_t rank_id, int64_t group_size)
-{
-    peer_cache_entry_t *entry = GET_GROUP_RANK_CACHE_ENTRY(cache, gp_uid, rank_id, group_size);
-    if (entry == NULL)
-        return false;
-    return (entry->set);
-}
 
 dpu_offload_status_t do_send_add_group_rank_request(execution_context_t *econtext, ucp_ep_h ep, uint64_t dest_id, dpu_offload_event_t *ev)
 {
@@ -264,43 +255,6 @@ dpu_offload_status_t get_local_sp_id(offloading_engine_t *engine, uint64_t globa
         return DO_ERROR;
     *local_sp_id = sp_info->service_proc.local_id;
     return DO_SUCCESS;
-}
-
-uint64_t LOCAL_ID_TO_GLOBAL(execution_context_t *econtext, uint64_t local_id)
-{
-    uint64_t global_id = UINT64_MAX;
-    switch (econtext->type)
-    {
-    case CONTEXT_SERVER:
-    {
-        peer_info_t *_c = DYN_ARRAY_GET_ELT(&(econtext->server->connected_clients.clients), local_id, peer_info_t);
-        assert(_c);
-        global_id = _c->rank_data.group_rank;
-        break;
-    }
-    case CONTEXT_CLIENT:
-    {
-        if (econtext->engine->on_dpu)
-            global_id = local_id;
-        else
-            global_id = econtext->client->server_global_id;
-        break;
-    }
-    case CONTEXT_SELF:
-    {
-        if (econtext->engine->on_dpu)
-            global_id = econtext->engine->config->local_service_proc.info.global_id;
-        else
-            global_id = 0; /* By default, for comm to self the ID is 0 */
-        break;
-    }
-    default:
-    {
-        global_id = UINT64_MAX;
-        break;
-    }
-    }
-    return global_id;
 }
 
 dpu_offload_status_t do_send_cache_entry_request(execution_context_t *econtext, ucp_ep_h ep, uint64_t dest_id, rank_info_t *requested_peer, dpu_offload_event_t *ev)
@@ -818,32 +772,6 @@ dpu_offload_status_t broadcast_group_cache(offloading_engine_t *engine, group_ui
     return DO_SUCCESS;
 }
 
-void display_group_cache(cache_t *cache, group_uid_t gp_uid)
-{
-    size_t i = 0;
-    size_t idx = 0;
-    group_cache_t *gp_cache = NULL;
-    gp_cache = GET_GROUP_CACHE(cache, gp_uid);
-    fprintf(stderr, "Content of cache for group 0x%x\n", gp_uid);
-    fprintf(stderr, "-> group size: %ld\n", gp_cache->group_size);
-    fprintf(stderr, "-> n_local_rank: %ld\n", gp_cache->n_local_ranks);
-    fprintf(stderr, "-> n_local_ranks_populated: %ld\n", gp_cache->n_local_ranks_populated);
-    fprintf(stderr, "-> num_local_entries: %ld\n", gp_cache->num_local_entries);
-    fprintf(stderr, "-> sent_to_host: %d\n\n", gp_cache->sent_to_host);
-
-    while (i < gp_cache->group_size)
-    {
-        peer_cache_entry_t *entry = GET_GROUPRANK_CACHE_ENTRY(cache, gp_uid, idx);
-        if (entry->set)
-        {
-            fprintf(stderr, "Rank %" PRId64 " host: 0x%lx\n", entry->peer.proc_info.group_rank, entry->peer.host_info);
-            assert(idx == entry->peer.proc_info.group_rank);
-            i++;
-        }
-        idx++;
-    }
-}
-
 dpu_offload_status_t get_group_rank_host(offloading_engine_t *engine,
                                          group_uid_t gp_uid,
                                          int64_t rank,
@@ -984,177 +912,6 @@ dpu_offload_status_t get_group_rank_sps(offloading_engine_t *engine,
     }
     *n_sps = num;
     return DO_SUCCESS;
-}
-
-static dpu_offload_status_t
-do_get_cache_entry_by_group_rank(offloading_engine_t *engine,
-                                 group_uid_t gp_uid,
-                                 int64_t rank,
-                                 int64_t sp_idx,
-                                 request_compl_cb_t cb,
-                                 int64_t *sp_global_id,
-                                 dpu_offload_event_t **ev)
-{
-    if (ev != NULL && cb != NULL)
-    {
-        ERR_MSG("%s(): both the event and the callback are defined, impossible to understand the context", __func__);
-        return DO_ERROR;
-    }
-
-    // If the event is defined, the dpu_id must also be defined, they go in pairs
-    if (ev != NULL)
-        assert(sp_global_id);
-
-    if (is_in_cache(&(engine->procs_cache), gp_uid, rank, -1))
-    {
-        // The cache has the data
-        peer_cache_entry_t *cache_entry = GET_GROUP_RANK_CACHE_ENTRY(&(engine->procs_cache), gp_uid, rank, GROUP_SIZE_UNKNOWN);
-        DBG("%" PRId64 " from group 0x%x is in the cache, service proc ID = %" PRId64, rank, gp_uid, cache_entry->shadow_service_procs[sp_idx]);
-        if (ev != NULL)
-        {
-            *ev = NULL;
-            *sp_global_id = cache_entry->shadow_service_procs[sp_idx];
-        }
-        return DO_SUCCESS;
-    }
-
-#if !NDEBUG
-    // With the current implementation, it should always be in the cache
-    WARN_MSG("rank %" PRId64 " from group 0x%x is not in the cache", rank, gp_uid);
-    display_group_cache(&(engine->procs_cache), gp_uid);
-    assert(0);
-#endif
-
-    // If the element is not in the cache, we send a request to the service process to get the data.
-    // At the moment we assume by design that the cache is fully populated early on so we should never
-    // face this situation. We keep the code in case we relax the assumptions later so we won't have
-    // to start from scratch.
-    abort();
-
-    // The cache does not have the data. We sent a request to get the data.
-    // The caller is in charge of calling the function after completion to actually get the data
-    rank_info_t rank_data;
-    RESET_RANK_INFO(&rank_data);
-    rank_data.group_uid = gp_uid;
-    rank_data.group_rank = rank;
-
-    // Create the local event so we can know when the cache entry has been received
-    dpu_offload_event_t *cache_entry_updated_ev;
-    peer_cache_entry_t *cache_entry = GET_GROUP_RANK_CACHE_ENTRY(&(engine->procs_cache), gp_uid, rank, GROUP_SIZE_UNKNOWN);
-    dpu_offload_status_t rc = event_get(engine->self_econtext->event_channels, NULL, &cache_entry_updated_ev);
-    CHECK_ERR_RETURN((rc), DO_ERROR, "event_get() failed");
-    if (!cache_entry->events_initialized)
-    {
-        SIMPLE_LIST_INIT(&(cache_entry->events));
-        cache_entry->events_initialized = true;
-    }
-    EVENT_HDR_TYPE(cache_entry_updated_ev) = META_EVENT_TYPE;
-    // We just queue a local event on the list for the cache entry to track what is being done in the
-    // context of that entry
-    SIMPLE_LIST_PREPEND(&(cache_entry->events), &(cache_entry_updated_ev->item));
-    DBG("Cache entry %p for gp/rank 0x%x/%" PRIu64 " now has %ld update events",
-        cache_entry, gp_uid, rank, SIMPLE_LIST_LENGTH(&(cache_entry->events)));
-    if (ev != NULL)
-    {
-        // If the calling function is expecting an event back, no need for anything other than
-        // make sure we return the event
-        *ev = cache_entry_updated_ev;
-    }
-    if (cb != NULL)
-    {
-        // If the calling function specified a callback, the event needs to be hidden from the
-        // caller and the callback will be invoked upon completion. So we need to put the event
-        // on the list of ongoing events. Make sure to never return the event to the caller to
-        // prevent the case where the event could be put on two different lists (which leads
-        // to memory corruptions).
-        cache_entry_request_t *request_data;
-        DYN_LIST_GET(engine->free_cache_entry_requests, cache_entry_request_t, item, request_data);
-        assert(request_data);
-        request_data->gp_uid = gp_uid;
-        request_data->rank = rank;
-        request_data->target_sp_idx = sp_idx;
-        request_data->offload_engine = engine;
-        cache_entry_updated_ev->ctx.completion_cb = cb;
-        cache_entry_updated_ev->ctx.completion_cb_ctx = (void *)request_data;
-        assert(0); // FIXME: events cannot be on two lists
-        // ucs_list_add_tail(&(engine->self_econtext->ongoing_events), &(cache_entry_updated_ev->item));
-    }
-
-    if (engine->on_dpu == true)
-    {
-        // If we are on a DPU, we need to send a request to all known DPUs
-        // To track completion, we get an event from the execution context used for the
-        // first DPU.
-        size_t i;
-        dpu_offload_status_t rc;
-        dpu_offload_event_t *metaev = NULL;
-        execution_context_t *meta_econtext = NULL;
-
-        for (i = 0; i < engine->num_service_procs; i++)
-        {
-            remote_service_proc_info_t *sp;
-            sp = DYN_ARRAY_GET_ELT(&(engine->service_procs), i, remote_service_proc_info_t);
-            assert(sp);
-            if (sp != NULL && sp->ep != NULL && sp->init_params.conn_params != NULL)
-            {
-                execution_context_t *econtext = ECONTEXT_FOR_SERVICE_PROC_COMMUNICATION(engine, i);
-                CHECK_ERR_RETURN((econtext == NULL), DO_ERROR, "unable to get execution context to communicate with service process #%ld", i);
-                uint64_t global_sp_id = LOCAL_ID_TO_GLOBAL(econtext, i);
-                DBG("Sending cache entry request to service process #%ld (econtext: %p, scope_id: %d)",
-                    global_sp_id,
-                    econtext,
-                    econtext->scope_id);
-
-                if (metaev == NULL)
-                {
-                    meta_econtext = econtext;
-                    rc = event_get(meta_econtext->event_channels, NULL, &metaev);
-                    CHECK_ERR_RETURN((rc), DO_ERROR, "get_event() failed");
-                    EVENT_HDR_TYPE(metaev) = META_EVENT_TYPE;
-                }
-
-                ucp_ep_h dpu_ep = sp->ep;
-                dpu_offload_event_t *subev;
-                rc = event_get(econtext->event_channels, NULL, &subev);
-                CHECK_ERR_RETURN((rc), DO_ERROR, "event_get() failed");
-                subev->is_subevent = true;
-                rc = do_send_cache_entry_request(econtext, dpu_ep, i, &rank_data, subev);
-                CHECK_ERR_RETURN((rc), DO_ERROR, "send_cache_entry_request() failed");
-                DBG("Sub-event for sending cache to DPU %ld: %p", global_sp_id, subev);
-                if (subev != NULL)
-                {
-                    // If the event did not complete right away, we add it as a sub-event to the meta-event so we can track everything
-                    QUEUE_SUBEVENT(metaev, subev);
-                }
-            }
-        }
-        if (metaev)
-        {
-            assert(meta_econtext);
-            if (!event_completed(metaev))
-                QUEUE_EVENT(metaev);
-            else
-                event_return(&metaev);
-        }
-    }
-    else
-    {
-        // If we are on the host, we need to send a request to our first shadow DPU
-        DBG("Sending request for cache entry...");
-        execution_context_t *econtext = engine->client;
-        return send_cache_entry_request(econtext, GET_SERVER_EP(econtext), econtext->client->server_id, &rank_data, ev);
-    }
-    return DO_SUCCESS;
-}
-
-dpu_offload_status_t get_cache_entry_by_group_rank(offloading_engine_t *engine, group_uid_t gp_uid, int64_t rank, int64_t sp_idx, request_compl_cb_t cb)
-{
-    return do_get_cache_entry_by_group_rank(engine, gp_uid, rank, sp_idx, cb, NULL, NULL);
-}
-
-dpu_offload_status_t get_sp_id_by_group_rank(offloading_engine_t *engine, group_uid_t gp_uid, int64_t rank, int64_t sp_idx, int64_t *sp_id, dpu_offload_event_t **ev)
-{
-    return do_get_cache_entry_by_group_rank(engine, gp_uid, rank, sp_idx, NULL, sp_id, ev);
 }
 
 dpu_offload_status_t get_sp_ep_by_id(offloading_engine_t *engine, uint64_t sp_id, ucp_ep_h *sp_ep, execution_context_t **econtext_comm, uint64_t *comm_id)
